+++ src/main/java/kitchenpos/domain/order/OrderService.java
@@ -1,9 +1,9 @@
 package kitchenpos.domain.order;
 
-import kitchenpos.application.dto.request.CreateOrderRequest;
-import kitchenpos.application.dto.request.UpdateOrderStatusRequest;
-import kitchenpos.application.dto.response.CreateOrderResponse;
-import kitchenpos.application.dto.response.OrderResponse;
+import kitchenpos.dto.request.CreateOrderRequest;
+import kitchenpos.dto.request.UpdateOrderStatusRequest;
+import kitchenpos.dto.response.CreateOrderResponse;
+import kitchenpos.dto.response.OrderResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
