+++ src/main/java/kitchenpos/domain/product/ProductService.java
@@ -1,8 +1,8 @@
 package kitchenpos.domain.product;
 
-import kitchenpos.application.dto.request.CreateProductRequest;
-import kitchenpos.application.dto.response.CreateProductResponse;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.dto.request.CreateProductRequest;
+import kitchenpos.dto.response.CreateProductResponse;
+import kitchenpos.dto.response.ProductResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
