+++ src/main/java/kitchenpos/domain/tablegroup/TableGroupMapper.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.tablegroup;
 
-import kitchenpos.application.dto.request.CreateTableGroupRequest;
+import kitchenpos.dto.request.CreateTableGroupRequest;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
 import org.springframework.stereotype.Component;
