+++ src/main/java/kitchenpos/dto/response/CreateOrderResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.response;
+package kitchenpos.dto.response;
 
 import kitchenpos.domain.order.Order;
 
@@ -6,22 +6,22 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class OrderResponse {
+public class CreateOrderResponse {
     private final Long id;
     private final Long orderTableId;
     private final String orderStatus;
-    private final String orderedTime;
+    private final LocalDateTime orderedTime;
     private final List<OrderLineItemResponse> orderLineItems;
 
-    private OrderResponse(Long id, Long orderTableId, String orderStatus, LocalDateTime orderedTime, List<OrderLineItemResponse> orderLineItems) {
+    private CreateOrderResponse(Long id, Long orderTableId, String orderStatus, LocalDateTime orderedTime, List<OrderLineItemResponse> orderLineItems) {
         this.id = id;
         this.orderTableId = orderTableId;
         this.orderStatus = orderStatus;
-        this.orderedTime = orderedTime.toString();
+        this.orderedTime = orderedTime;
         this.orderLineItems = orderLineItems;
     }
 
-    public static OrderResponse from(Order order) {
+    public static CreateOrderResponse from(Order order) {
         return builder()
                 .id(order.getId())
                 .orderTableId(order.getOrderTableId())
@@ -33,47 +33,47 @@ public class OrderResponse {
                 .build();
     }
 
-    public static OrderResponseBuilder builder() {
-        return new OrderResponseBuilder();
+    public static CreateOrderResponseBuilder builder() {
+        return new CreateOrderResponseBuilder();
     }
 
-    public static final class OrderResponseBuilder {
+    public static final class CreateOrderResponseBuilder {
         private Long id;
         private Long orderTableId;
         private String orderStatus;
         private LocalDateTime orderedTime;
         private List<OrderLineItemResponse> orderLineItems;
 
-        private OrderResponseBuilder() {
+        private CreateOrderResponseBuilder() {
         }
 
-        public OrderResponseBuilder id(Long id) {
+        public CreateOrderResponseBuilder id(Long id) {
             this.id = id;
             return this;
         }
 
-        public OrderResponseBuilder orderTableId(Long orderTableId) {
+        public CreateOrderResponseBuilder orderTableId(Long orderTableId) {
             this.orderTableId = orderTableId;
             return this;
         }
 
-        public OrderResponseBuilder orderStatus(String orderStatus) {
+        public CreateOrderResponseBuilder orderStatus(String orderStatus) {
             this.orderStatus = orderStatus;
             return this;
         }
 
-        public OrderResponseBuilder orderedTime(LocalDateTime orderedTime) {
+        public CreateOrderResponseBuilder orderedTime(LocalDateTime orderedTime) {
             this.orderedTime = orderedTime;
             return this;
         }
 
-        public OrderResponseBuilder orderLineItems(List<OrderLineItemResponse> orderLineItems) {
+        public CreateOrderResponseBuilder orderLineItems(List<OrderLineItemResponse> orderLineItems) {
             this.orderLineItems = orderLineItems;
             return this;
         }
 
-        public OrderResponse build() {
-            return new OrderResponse(id, orderTableId, orderStatus, orderedTime, orderLineItems);
+        public CreateOrderResponse build() {
+            return new CreateOrderResponse(id, orderTableId, orderStatus, orderedTime, orderLineItems);
         }
     }
 
@@ -89,7 +89,7 @@ public class OrderResponse {
         return orderStatus;
     }
 
-    public String getOrderedTime() {
+    public LocalDateTime getOrderedTime() {
         return orderedTime;
     }
