+++ src/main/java/kitchenpos/dto/response/CreateTableGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto.response;
+package kitchenpos.dto.response;
 
 import kitchenpos.domain.tablegroup.TableGroup;
