+++ src/main/java/kitchenpos/dto/response/ProductResponse.java
@@ -1,55 +1,55 @@
-package kitchenpos.application.dto.response;
+package kitchenpos.dto.response;
 
 import kitchenpos.domain.product.Product;
 
-public class CreateProductResponse {
+public class ProductResponse {
     private final Long id;
     private final String name;
     private final String price;
 
-    private CreateProductResponse(Long id, String name, String price) {
+    private ProductResponse(Long id, String name, String price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public static CreateProductResponse from(final Product product) {
-        return CreateProductResponse.builder()
+    public static ProductResponse from(final Product product) {
+        return ProductResponse.builder()
                 .id(product.getId())
                 .name(product.getName())
                 .price(product.getPrice().toString())
                 .build();
     }
 
-    public static CreateProductResponseBuilder builder() {
-        return new CreateProductResponseBuilder();
+    public static ProductResponseBuilder builder() {
+        return new ProductResponseBuilder();
     }
 
-    public static final class CreateProductResponseBuilder {
+    public static final class ProductResponseBuilder {
         private Long id;
         private String name;
         private String price;
 
-        private CreateProductResponseBuilder() {
+        private ProductResponseBuilder() {
         }
 
-        public CreateProductResponseBuilder id(Long id) {
+        public ProductResponseBuilder id(Long id) {
             this.id = id;
             return this;
         }
 
-        public CreateProductResponseBuilder name(String name) {
+        public ProductResponseBuilder name(String name) {
             this.name = name;
             return this;
         }
 
-        public CreateProductResponseBuilder price(String price) {
+        public ProductResponseBuilder price(String price) {
             this.price = price;
             return this;
         }
 
-        public CreateProductResponse build() {
-            return new CreateProductResponse(id, name, price);
+        public ProductResponse build() {
+            return new ProductResponse(id, name, price);
         }
     }
