+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -1,9 +1,9 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.dto.request.CreateOrderRequest;
-import kitchenpos.application.dto.request.UpdateOrderStatusRequest;
-import kitchenpos.application.dto.response.CreateOrderResponse;
-import kitchenpos.application.dto.response.OrderResponse;
+import kitchenpos.dto.request.CreateOrderRequest;
+import kitchenpos.dto.request.UpdateOrderStatusRequest;
+import kitchenpos.dto.response.CreateOrderResponse;
+import kitchenpos.dto.response.OrderResponse;
 import kitchenpos.domain.order.OrderService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
