+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.dto.request.CreateProductRequest;
-import kitchenpos.application.dto.response.CreateProductResponse;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.dto.request.CreateProductRequest;
+import kitchenpos.dto.response.CreateProductResponse;
+import kitchenpos.dto.response.ProductResponse;
 import kitchenpos.domain.product.ProductService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
