+++ src/test/java/kitchenpos/acceptance/OrderTableTest.java
@@ -5,8 +5,8 @@ import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.Acceptance;
-import kitchenpos.application.dto.request.CreateOrderTableRequest;
-import kitchenpos.application.dto.response.OrderTableResponse;
+import kitchenpos.dto.request.CreateOrderTableRequest;
+import kitchenpos.dto.response.OrderTableResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
