+++ src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
-import kitchenpos.application.dto.request.MenuProductRequest;
-import kitchenpos.application.dto.response.MenuProductResponse;
+import kitchenpos.dto.request.MenuProductRequest;
+import kitchenpos.dto.response.MenuProductResponse;
 import kitchenpos.domain.menu.MenuProduct;
 
 import static kitchenpos.fixture.ProductFixture.PRODUCT;
